chore: community navigation 구조 수정 및 화면 추가
