+++ src/navigations/topTab/CommunityTopTabNavigator.tsx
@@ -22,10 +22,10 @@ function CommunityTopTabNavigator() {
 				screenOptions={{
 					tabBarLabelStyle: {
 						fontSize: 20,
-						fontFamily: 'Pretendard-Bold',
+						fontFamily: 'Pretendard-Light',
 					},
 					tabBarInactiveTintColor: colors[theme].GRAY_500,
-					tabBarActiveTintColor: colors[theme].PINK_700,
+					tabBarActiveTintColor: colors[theme].BLACK,
 					tabBarStyle: {
 						backgroundColor: colors[theme].WHITE,
 					},
